asset_tracker/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-filter==3.1.0